refactor(morpho-sdk): align midnight entity type layout
