+++ packages/morpho-sdk/src/entities/midnight/index.ts
@@ -1,1 +1,2 @@
 export * from "./midnight.js";
+export * from "./types.js";
